feat: 캐치마인드 컨트롤러 분리 및 커넥터 연결
